requirements.txt
Flask~=3.0.0
Faker~=20.0.3
werkzeug~=3.0.1
google~=3.0.0
python-dotenv~=1.0.0